requirements.txt
Django==2.2.14
djangorestframework>=3.9.0,<3.10.0
flake8>=3.6.0,<3.7.0
django-extensions==2.1.6
ipdb==0.13.3